task2/apiservice/requirements.txt
Django==4.1.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
frozenlist==1.3.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
multidict==6.0.2
numpy==1.23.4
pandas==1.5.0
python-dateutil==2.8.2
pytz==2022.4
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
psycopg2-binary
requests
openpyxl
et-xmlfile
